added the HasLocation mixin and updated data model
